requirements.txt
ffmpeg-python
openai-whisper
torch
transformers
ultralytics
easyocr
opencv-python
numpy
pandas
tqdm


